Upg: split notion workflow in 2 and reorg files
