+++ connectors/src/start.ts
@@ -8,7 +8,10 @@ import { runSnowflakeWorker } from "@connectors/connectors/snowflake/temporal/wo
 import { runGithubWorker } from "./connectors/github/temporal/worker";
 import { runGoogleWorkers } from "./connectors/google_drive/temporal/worker";
 import { runIntercomWorker } from "./connectors/intercom/temporal/worker";
-import { runNotionWorker } from "./connectors/notion/temporal/worker";
+import {
+  runNotionGarbageCollectWorker,
+  runNotionWorker,
+} from "./connectors/notion/temporal/worker";
 import { runSlackWorker } from "./connectors/slack/temporal/worker";
 import { runWebCrawlerWorker } from "./connectors/webcrawler/temporal/worker";
 import { errorFromAny } from "./lib/error";
@@ -30,6 +33,9 @@ runSlackWorker().catch((err) =>
 );
 runNotionWorker().catch((err) =>
   logger.error(errorFromAny(err), "Error running notion worker")
+);
+runNotionGarbageCollectWorker().catch((err) =>
+  logger.error(errorFromAny(err), "Error running notion gc worker")
 );
 runGithubWorker().catch((err) =>
   logger.error(errorFromAny(err), "Error running github worker")
